Add UI entry point in manifest, update package scripts for testing, and enhance TypeScript configuration
