+++ src/ui.ts
@@ -1,0 +1,5 @@
+const App = (function App() {
+	return `<p>For testing purposes only</p>`
+})()
+
+document.querySelector('#app').innerHTML = App
