+++ tsconfig.json
@@ -1,5 +1,21 @@
 {
 	"compilerOptions": {
-		"typeRoots": ["node_modules/@figma"]
-	}
+		"target": "ESNext",
+		"useDefineForClassFields": true,
+		"module": "nodenext",
+        "moduleResolution": "nodenext",
+		/**
+		 * Typecheck JS in `.svelte` and `.js` files by default.
+		 * Disable checkJs if you'd like to use dynamic types in JS.
+		 * Note that setting allowJs false does not prevent the use
+		 * of JS in `.svelte` files.
+		 */
+		"allowJs": true,
+		"checkJs": true,
+		"isolatedModules": true,
+		"moduleDetection": "force",
+		"typeRoots": ["node_modules/@figma", "node_modules/@types"],
+	},
+	"include": ["src/**/*.ts", "src/**/*.js", "e2e/**/*.ts", "tests/**/*.ts"],
+	"exclude": ["node_modules"]
 }
